# public/openapi.yaml
openapi: 3.1.0
info:
  title: CodeStride API
  description: REST API for CodeStride - an open-source developer productivity dashboard
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://codestride.vercel.app
    description: Production server
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: next-auth.session-token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...
paths:
...
  /api/auth/link-github:
    get:
      tags: [Auth]
      summary: Initiate GitHub account linking
      security:
        - cookieAuth: []
      responses:
        '302':
          description: Redirects to GitHub OAuth
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/link-github/callback:
    get:
      tags: [Auth]
      summary: GitHub OAuth callback for account linking
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        '302':
          description: Redirects after successful linking
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # BADGE ROUTES
...
  /api/leaderboard:
    get:
      tags: [Leaderboard]
      summary: Get public leaderboard rankings
      parameters:
        - name: lang
          in: query
          required: false
          schema:
            type: string
          description: Filter by primary language
        - name: period
          in: query
          required: false
          schema:
            type: string
            enum: [week, month, year]
          description: Filter by time period
      responses:
        '200':
          description: Leaderboard data
          content:
            application/json:
              schema:
                type: object
        '429':
          description: Rate limit exceeded
...
  /api/leaderboard/rebuild:
    post:
      tags: [Leaderboard]
      summary: Full leaderboard rebuild
      security:
        - bearerAuth: []
      parameters:
        - name: Authorization
          in: header
          required: true
          schema:
            type: string
          description: Bearer token with LEADERBOARD_REBUILD_TOKEN
      responses:
        '200':
          description: Leaderboard rebuild started
        '401':
          description: Invalid or missing rebuild token

  # PUBLIC PROFILE ROUTES
...